[C++] Generate a static <table>::Create method.
This mirrors the signature, implementation, and usage of the "CreateX"
functions that already exist, but have the advantage that they are
accessible given the type name of the generated class and do not
require a specialized (and hardcoded) identifier in user code.

Tested by running unit tests and examining generated code.

Fixes #5559